moved: src/components/... to atomic design folders added: preprocessorOptions > variables import for all scss files
