+++ src/App.scss
@@ -1,5 +1,3 @@
-@use './styles/variables' as *;
-
 .app {
     &__footer {
         margin: $spacing-medium 0;
