+++ src/App.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import type { JSX } from 'react';
 import { HashRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.scss';
-import LinksSection from '@/components/LinksSection';
-import Profile from '@/components/Profile';
+import LinksSection from '@/components/molecules/LinksSection';
+import Profile from '@/components/organisms/Profile';
 import Schedules from '@/pages/Schedules';
 import Castings from '@/pages/Castings';
 import Workshops from '@/pages/Workshops';
-import ScrollToTop from '@/components/ScrollToTop';
+import ScrollToTop from '@/components/molecules/ScrollToTop';
 import { linksSections } from '@/data/linksSection';
 import type { LinksSectionsProps } from '@/types';
