+++ src/components/Atoms/Buttons/Button.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .button {
   display: inline-block;
   align-content: center;
