+++ src/components/Atoms/Buttons/RadioButton.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .radio-button {
   position: relative;
   display: inline-block;
