+++ src/components/Molecules/LinksSection/LinksSection.scss
@@ -1,5 +1,3 @@
-@use '../../styles/variables' as *;
-
 // Componente de sección de enlaces
 .links-section {
   // Box Model
