+++ src/components/Molecules/Modal/Modal.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables.scss' as *;
-
 .modal {
   position: fixed;
   top: 0;
