+++ src/components/Molecules/WhatsAppButton/WhatsAppButton.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .whatsapp-button {
   // Personalizaciones específicas del botón de WhatsApp
   // La mayoría de los estilos vienen del componente Button base
