+++ src/components/organisms/Castings/CastingList.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .castings-container {
 
   .casting-card {
