+++ src/components/organisms/Castings/GenderFilter.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .gender-filter {
   margin-bottom: $spacing-medium;
