+++ src/components/organisms/Navbar/Navbar.scss
@@ -1,5 +1,3 @@
-@use '../../styles/variables' as *;
-
 // Componente principal de navegación
 .navbar {
   // Positioning
