+++ src/components/organisms/Profile/Profile.scss
@@ -1,5 +1,3 @@
-@use '../../styles/variables' as *;
-
 // Componente de perfil principal
 .profile {
   // Box Model
