+++ src/components/organisms/Schedule/Schedule.scss
@@ -1,5 +1,3 @@
-@use '../../styles/variables' as *;
-
 .schedule {
   &__content {
     box-shadow: 0 4px 6px $shadow-soft;
