+++ src/components/organisms/Workshops/WorkshopList.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .workshops-container {
 
   .filter-buttons {
