+++ src/pages/Castings.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .castings {
   &__container {
     // Positioning
