+++ src/pages/Schedules.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .schedules {
   &__selector {
     display: flex;
