+++ src/pages/Workshops.scss
@@ -1,5 +1,3 @@
-@use '@/styles/variables' as *;
-
 .workshops-description {
   margin-bottom: $spacing-large;
 }
