+++ src/styles/_fonts.scss
@@ -1,5 +1,3 @@
-@use 'variables' as *;
-
 // Brice Font
 @font-face {
   font-family: 'Brice';
