+++ src/styles/main.scss
@@ -1,4 +1,3 @@
-@use 'variables' as *;
 @use 'fonts';
 
 // Reset básico
