+++ vite.config.ts
@@ -14,6 +14,7 @@ export default defineConfig({
   css: {
     preprocessorOptions: {
       scss: {
+        additionalData: `@use "@/styles/_variables" as *;`,
       }
     }
   }
